[ISSUE-27] Refactored Logo component
